** removed more useless conversions
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@6020 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/FeatureFinder_impl.h
@@ -103,7 +103,7 @@ namespace OpenMS
 		{
 			//index
 			Size spectrum_index = input_map.RTBegin(features[i].getRT()) - input_map.begin();
-			features[i].setMetaValue("spectrum_index",(UInt) spectrum_index);
+			features[i].setMetaValue("spectrum_index", spectrum_index);
 			//native id
 			if (spectrum_index < input_map.size())
 			{
